+++ lib/td_auth/accounts/accounts.ex
@@ -14,8 +14,8 @@ defmodule TdAuth.Accounts do
 
   def user_exists? do
     User
-    |> where([_u], is_protected: false)
-    |> Repo.exists?()
+    |> where([u], u.role == :admin)
+    |> Repo.exists?(role: :admin)
   end
 
   @doc """
@@ -28,16 +28,14 @@ defmodule TdAuth.Accounts do
 
   """
   def list_users(opts \\ []) do
-    filter_clauses = Keyword.put_new(opts, :is_protected, false)
-
     User
-    |> do_where(filter_clauses)
+    |> do_where(opts)
     |> Repo.all()
   end
 
   defp do_where(queryable, filter_clauses) do
     Enum.reduce(filter_clauses, queryable, fn
-      {:is_protected, is_protected}, q -> where(q, is_protected: ^is_protected)
+      {:role, role}, q -> where(q, role: ^role)
       {:id, {:in, ids}}, q -> where(q, [u], u.id in ^ids)
       {:preload, preloads}, q -> preload(q, ^preloads)
       _, q -> q